+++ RELEASE-NOTES.md
@@ -1,4 +1,6 @@
 ## RELEASE NOTES
+### Version 2.79.1-default-organisation-pre-release
+**EUI-3348** pre-select default organisation depends on PrepopulateToUsersOrganisation
 
 ### Version 2.79.1
 **EUI-3452** Fix to some issues on the Check your answers page.